requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
mlflow
pytest
transformers
datasets
sentencepiece
lightgbm
packaging>=20.9
zipp>=3.1.0
jupyterlab
jupyter-client>=6.1.5
traitlets>=5.0.0
pytorch-lightning
typing-extensions>=3.7.4.3
setuptools==59.5.0
tweepy
requests==2.27.0
urllib3>=1.25.10
python-dotenv